test(sync): move code
