+++ crates/papyrus_p2p_sync/src/header_test.rs
@@ -1,0 +1,163 @@
+use std::time::Duration;
+
+use futures::channel::mpsc::{Receiver, Sender};
+use futures::future::ready;
+use futures::{FutureExt, SinkExt, StreamExt};
+use indexmap::indexmap;
+use lazy_static::lazy_static;
+use papyrus_network::{DataType, Direction, Query, ResponseReceivers, SignedBlockHeader};
+use papyrus_storage::header::HeaderStorageReader;
+use papyrus_storage::state::StateStorageReader;
+use papyrus_storage::test_utils::get_test_storage;
+use papyrus_storage::StorageReader;
+use rand::RngCore;
+use starknet_api::block::{BlockHash, BlockHeader, BlockNumber, BlockSignature};
+use starknet_api::core::{ClassHash, CompiledClassHash, ContractAddress, Nonce};
+use starknet_api::crypto::Signature;
+use starknet_api::hash::{StarkFelt, StarkHash};
+use starknet_api::state::ThinStateDiff;
+use static_assertions::const_assert;
+use test_utils::get_rng;
+use tokio::time::timeout;
+
+use super::{P2PSync, P2PSyncConfig};
+
+#[tokio::test]
+async fn signed_headers_basic_flow() {
+    const NUM_QUERIES: usize = 3;
+
+    let (p2p_sync, storage_reader, query_receiver, mut signed_headers_sender, _state_diffs_sender) =
+        setup();
+    let block_hashes_and_signatures =
+        create_block_hashes_and_signatures((NUM_QUERIES * HEADER_QUERY_LENGTH).try_into().unwrap());
+
+    let mut query_receiver = query_receiver
+        .filter(|query| ready(matches!(query.data_type, DataType::SignedBlockHeader)));
+
+    // Create a future that will receive queries, send responses and validate the results.
+    let parse_queries_future = async move {
+        for query_index in 0..NUM_QUERIES {
+            let start_block_number = query_index * HEADER_QUERY_LENGTH;
+            let end_block_number = (query_index + 1) * HEADER_QUERY_LENGTH;
+
+            // Receive query and validate it.
+            let query = query_receiver.next().await.unwrap();
+            assert_eq!(
+                query,
+                Query {
+                    start_block: BlockNumber(start_block_number.try_into().unwrap()),
+                    direction: Direction::Forward,
+                    limit: HEADER_QUERY_LENGTH,
+                    step: 1,
+                    data_type: DataType::SignedBlockHeader,
+                }
+            );
+
+            for (i, (block_hash, block_signature)) in block_hashes_and_signatures
+                .iter()
+                .enumerate()
+                .take(end_block_number)
+                .skip(start_block_number)
+            {
+                // Send responses
+                signed_headers_sender
+                    .send(Some(SignedBlockHeader {
+                        block_header: BlockHeader {
+                            block_number: BlockNumber(i.try_into().unwrap()),
+                            block_hash: *block_hash,
+                            state_diff_length: Some(0),
+                            ..Default::default()
+                        },
+                        signatures: vec![*block_signature],
+                    }))
+                    .await
+                    .unwrap();
+
+                tokio::time::sleep(SLEEP_DURATION_TO_LET_SYNC_ADVANCE).await;
+
+                // Check responses were written to the storage. This way we make sure that the sync
+                // writes to the storage each response it receives before all query responses were
+                // sent.
+                let block_number = BlockNumber(i.try_into().unwrap());
+                let txn = storage_reader.begin_ro_txn().unwrap();
+                assert_eq!(block_number.unchecked_next(), txn.get_header_marker().unwrap());
+                let block_header = txn.get_block_header(block_number).unwrap().unwrap();
+                assert_eq!(block_number, block_header.block_number);
+                assert_eq!(*block_hash, block_header.block_hash);
+                let actual_block_signature =
+                    txn.get_block_signature(block_number).unwrap().unwrap();
+                assert_eq!(*block_signature, actual_block_signature);
+            }
+            signed_headers_sender.send(None).await.unwrap();
+        }
+    };
+
+    tokio::select! {
+        sync_result = p2p_sync.run() => {
+            sync_result.unwrap();
+            panic!("P2P sync aborted with no failure.");
+        }
+        _ = parse_queries_future => {}
+    }
+}
+
+#[tokio::test]
+async fn sync_sends_new_header_query_if_it_got_partial_responses() {
+    const NUM_ACTUAL_RESPONSES: u8 = 2;
+    assert!(usize::from(NUM_ACTUAL_RESPONSES) < HEADER_QUERY_LENGTH);
+
+    let (p2p_sync, _storage_reader, query_receiver, mut signed_headers_sender, _state_diffs_sender) =
+        setup();
+    let block_hashes_and_signatures = create_block_hashes_and_signatures(NUM_ACTUAL_RESPONSES);
+
+    let mut query_receiver = query_receiver
+        .filter(|query| ready(matches!(query.data_type, DataType::SignedBlockHeader)));
+
+    // Create a future that will receive a query, send partial responses and receive the next query.
+    let parse_queries_future = async move {
+        let _query = query_receiver.next().await.unwrap();
+
+        for (i, (block_hash, signature)) in block_hashes_and_signatures.into_iter().enumerate() {
+            signed_headers_sender
+                .send(Some(SignedBlockHeader {
+                    block_header: BlockHeader {
+                        block_number: BlockNumber(i.try_into().unwrap()),
+                        block_hash,
+                        state_diff_length: Some(0),
+                        ..Default::default()
+                    },
+                    signatures: vec![signature],
+                }))
+                .await
+                .unwrap();
+        }
+        signed_headers_sender.send(None).await.unwrap();
+
+        // First unwrap is for the timeout. Second unwrap is for the Option returned from Stream.
+        let query = timeout(TIMEOUT_FOR_NEW_QUERY_AFTER_PARTIAL_RESPONSE, query_receiver.next())
+            .await
+            .unwrap()
+            .unwrap();
+
+        assert_eq!(
+            query,
+            Query {
+                start_block: BlockNumber(NUM_ACTUAL_RESPONSES.into()),
+                direction: Direction::Forward,
+                limit: HEADER_QUERY_LENGTH,
+                step: 1,
+                data_type: DataType::SignedBlockHeader,
+            }
+        );
+    };
+
+    tokio::select! {
+        sync_result = p2p_sync.run() => {
+            sync_result.unwrap();
+            panic!("P2P sync aborted with no failure.");
+        }
+        _ = parse_queries_future => {}
+    }
+}
+
+// TODO(shahak): Add negative tests.
